+++ scripts/run-test262-cli.ts
@@ -1,0 +1,3 @@
+import runTest from './run-test262';
+
+runTest();
