+++ test/test262-parser-tests/unsupported-features.txt
@@ -1,0 +1,5 @@
+decorators
+explicit-resource-management
+regexp-modifiers
+import-assertions
+import-attributes
